Place and format the current element of the page
Created many classes to display correctly the elements on the page
(Tweet button, header, tweet message box, tweet counter and all of
those together). Bordered all of those with rounded edges. Centered
the message box. Made a bottom border under the header. Put the tweet
button on the left and the tweet counter on the right, under the
message box.